Lower utility tests to the unit level
